+++ tests/basic_auth_test.rs
@@ -98,6 +98,7 @@ async fn public_basic_auth_rejects_and_allows() -> Result<()> {
             force_https: false,
             basic_auth: Some(CREDS.into()),
             notes: None,
+            ..Default::default()
         },
     )
     .await?;
@@ -152,6 +153,7 @@ async fn public_basic_auth_passes_non_http() -> Result<()> {
             force_https: false,
             basic_auth: Some(CREDS.into()),
             notes: None,
+            ..Default::default()
         },
     )
     .await?;